feat: use WyRng instead of Pcg32 for better aarch64 performance
