+++ src/utils/mod.rs
@@ -1,3 +1,3 @@
-pub mod pcg32;
 pub mod shim;
 pub mod unreachable;
+pub mod wyrng;
